requirements.txt
grpcio>=1.68.0
protobuf>=5.0.0
rpc-demo-stub @ git+https://github.com/SilverMelon233/rpc-stub.git#subdirectory=python
